+++ docs/svelte.config.js
@@ -5,16 +5,26 @@ const slug = require("remark-slug");
 const visit = require("unist-util-visit");
 const { format } = require("prettier");
 const pkg = require("../package.json");
+const component_api = require("./src/COMPONENT_API.json");
 const fs = require("fs");
 const Prism = require("prismjs");
 require("prism-svelte");
 
+const component_api_by_name = component_api.components.reduce((a, c) => {
+  return {
+    ...a,
+    [c.moduleName]: true,
+  };
+}, {});
+
 function createImports(source) {
   const inlineComponents = new Set();
   const icons = new Set();
+  const actions = new Set();
 
-  // TODO: [refactor] better determine if component is a Carbon icon
-  const isIcon = (text) => text.match(/(16|20|24|32)/);
+  // heuristic to guess if the inline component or expression name is a Carbon icon
+  const isIcon = (text) =>
+    /[A-Z][a-z]*/.test(text) && !(text in component_api_by_name);
 
   walk(parse(source), {
     enter(node) {
@@ -31,15 +41,20 @@ function createImports(source) {
         ) {
           icons.add(node.expression.name);
         }
+      } else if (node.type === "Action") {
+        actions.add(node.name);
       }
     },
   });
 
-  const ccs_imports = Array.from(inlineComponents.keys());
+  const action_imports = Array.from(actions.keys());
+  const ccs_imports = [
+    ...Array.from(inlineComponents.keys()),
+    ...action_imports,
+  ];
   const icon_imports = Array.from(icons.keys());
 
   if (ccs_imports.length === 0) return "";
-  // TODO: determine if action is used, and generate import accordingly
 
   return `
   <script>
@@ -48,7 +63,8 @@ function createImports(source) {
       icons.size > 0
         ? icon_imports
             .map(
-              (icon) => `import ${icon} from "carbon-icons-svelte/lib/${icon}";`
+              (icon) =>
+                `import ${icon} from "carbon-icons-svelte/lib/${icon}.svelte";`
             )
             .join("\n")
         : ""