+++ src/AspectRatio/AspectRatio.svelte
@@ -1,7 +1,7 @@
 <script>
   /**
    * Specify the aspect ratio
-   * @type {"2x1" | "16x9" | "4x3" | "1x1" | "3x4" | "3x2" | "9x16" | "1x2"}
+   * @type {"2x1" | "2x3" | "16x9" | "4x3" | "1x1" | "3x4" | "3x2" | "9x16" | "1x2"}
    */
   export let ratio = "2x1";
 </script>
@@ -9,6 +9,7 @@
 <div
   class:bx--aspect-ratio="{true}"
   class:bx--aspect-ratio--2x1="{ratio === '2x1'}"
+  class:bx--aspect-ratio--2x3="{ratio === '2x3'}"
   class:bx--aspect-ratio--16x9="{ratio === '16x9'}"
   class:bx--aspect-ratio--4x3="{ratio === '4x3'}"
   class:bx--aspect-ratio--1x1="{ratio === '1x1'}"