+++ src/ComboBox/ComboBox.svelte
@@ -76,10 +76,18 @@
   export let shouldFilterItem = () => true;
 
   /**
-   * Override the default translation ids
-   * @type {(id: any) => string}
+   * Override the chevron icon label based on the open state.
+   * Defaults to "Open menu" when closed and "Close menu" when open
+   * @type {(id: import("../ListBox/ListBoxMenuIcon.svelte").ListBoxMenuIconTranslationId) => string}
    */
   export let translateWithId = undefined;
+
+  /**
+   * Override the label of the clear button when the input has a selection.
+   * Defaults to "Clear selected item" since a combo box can only have on selection.
+   * @type {(id: "clearSelection") => string}
+   */
+  export let translateWithIdSelection = undefined;
 
   /** Set an id for the list box component */
   export let id = "ccs-" + Math.random().toString(36);
@@ -325,7 +333,7 @@
         <ListBoxSelection
           on:clear
           on:clear="{clear}"
-          translateWithId="{translateWithId}"
+          translateWithId="{translateWithIdSelection}"
           disabled="{disabled}"
           open="{open}"
         />