+++ src/DatePicker/DatePicker.svelte
@@ -60,11 +60,11 @@
   export let id = "ccs-" + Math.random().toString(36);
 
   /**
-   * Override the options passed to the Flatpickr instance
-   * https://flatpickr.js.org/options
+   * Override the options passed to the Flatpickr instance.
+   * @see https://flatpickr.js.org/options
    * @type {import("flatpickr/dist/types/options").Options}
    */
-  export let flatpickrProps = {};
+  export let flatpickrProps = { static: true };
 
   import {
     createEventDispatcher,
@@ -216,6 +216,9 @@
       locale,
       maxDate,
       minDate,
+      // default to static: true so the
+      // date picker works inside a modal
+      static: true,
       ...flatpickrProps,
     });
   }
@@ -252,14 +255,6 @@
       if (calendar?.isOpen && e.key === 'Escape') {
         e.stopPropagation();
         calendar.close();
-      }
-
-      if (
-        $hasCalendar &&
-        /INPUT/.test(document.activeElement?.tagName) &&
-        e.key === 'Enter'
-      ) {
-        e.stopPropagation();
       }
     }}"
   >