+++ types/Search/Search.svelte.d.ts
@@ -115,7 +115,7 @@ export default class Search extends SvelteComponentTyped<
     blur: WindowEventMap["blur"];
     keydown: WindowEventMap["keydown"];
     keyup: WindowEventMap["keyup"];
-    clear: CustomEvent<any>;
+    clear: CustomEvent<null>;
   },
   { labelText: {} }
 > {}